fix: add 404 page
This commit is based on the feedback from an instructor at Udacity:

You can also create a 404 error page so if someone goes to an incorrect page like http://localhost:3000/nothingshere it doesn't just show a blank page, but rather a 404 error.

To do this, you can create a component for when no pages match, then import Switch from react-router-dom and put your routes inside of opening and closing <Switch> tags so it will only render the first route that matches. You would then put the Route for when there is no match (your 404 / error page) as the last route.