fix: add missing cover image
based on feedback from an instructor at Udacity:

Another thing you could do is create a default cover image to be displayed if a book doesn't have a cover because you could put text on the default cover that says "No Cover" (or think of something better) so that people will know that book doesn't have a cover rather than wondering if it just didn't download the cover image properly.

You could test this out using this something like this as a default cover http://via.placeholder.com/128x193?text=No%20Cover (Use the URL in place of an image. You can customize the width, height, and text by changing the URL, but if you add very much text it will look too small at this image size.)

Try searching for 'biography' to test this as the book "My First Life" by Brenda Dean Paul that is normally in the results does not have the imageLinks property.